src/main/webapp/app/entities/abc-27/delete/abc-27-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { Abc27Service } from '../service/abc-27.service';

import { Abc27DeleteDialogComponent } from './abc-27-delete-dialog.component';

describe('Abc27 Management Delete Component', () => {
  let comp: Abc27DeleteDialogComponent;
  let fixture: ComponentFixture<Abc27DeleteDialogComponent>;
  let service: Abc27Service;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      declarations: [Abc27DeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(Abc27DeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(Abc27DeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(Abc27Service);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      })
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
